+++ README.md
@@ -18,9 +18,10 @@ made during university algorithms&alg.languages
 ### Hebi (по-японски "змейка"): стрелки, P=pause пауза, Q=quit выход
 
 * базовая змейка + pywin32/screeninfo + gfx для чуть красоты
-* минимализм как в дизайне, так и механике
+* минимализм как в дизайне, так и в механике
 * можно использовать как код для маскота (скопировать часть кода с pywin32/screeninfo и сделать своего персонажа, анимировать спрайты)
 * клавишное управление
+* выход в любое время Q
 * таблица рекордов подгружается с json
 * условно 4 экрана: старт, игра на дескотопе, конец игры, рекордный экран
 * помощь пользователю, есть инструкции какие клавишы использовать (стрелки, P=pause, Q=quit)
