+++ archive/archive_dist/snake_scores.json
@@ -1,0 +1,1 @@
+[{"name": "Excellent", "score": 20}]
