--- archive_dist/snake_scores.json
@@ -1,1 +1,0 @@
-[{"name": "Excellent", "score": 20}]
