+++ only HEBI GAME EXE/snake_scores.json
@@ -1,0 +1,1 @@
+[{"name": "ok", "score": 10}]
