--- snake_scores.json
@@ -1,1 +1,0 @@
-[{"name": "Pause", "score": 70}, {"name": "tt", "score": 60}, {"name": "wow", "score": 10}, {"name": "00", "score": 0}]
